run it on X86 system

+++ com/as.infrastructure/system/kernel/freeosek/src/posix/Osek_Internal_Arch.c
@@ -93,9 +93,9 @@ uint32 OsekHWTimer0;
 
 InterruptFlagsType InterruptFlag;
 
-uint32 PosixStack;
+unsigned long PosixStack;
 
-uint32 OsekStack;
+unsigned long OsekStack;
 
 /*==================[internal functions definition]==========================*/
